Fix Back method breaking when capacity and size were diferent.
When doing `Back` there is a corner case where we have to loop back into
the last position of the array. We must use capacity instead of size because
we want the internal buffer, last array index (which has nothing to do with
size if the array is not full).
Closes #2